requirements.txt
numpy
matplotlib
tensorflow-gpu
scikit-learn